ErrorHandler: call onException from onError to work around PHP #66216
parsing files including broken require_once (e.g. path can not be
resolved) will trigger an error which is handled by onError converting
it to an ErrorException and throwing it again.
This *should* then be caught by onException to convert and print this as
JSON and exit. But due to https://bugs.php.net/bug.php?id=66216 this
does not happen and the parser exits an empty output which results in
another Exception in php-proxy.coffee trying to parse the empty string
as JSON.